Don't echo the Datadog API key
This sends the API key in plaintext to Buildkite. This can be easily
seen by inspecting the build log of this plugin in the `Sending
deployment events to datadog` section.

Any user of version v0.0.4 should consider their Datadog API key
compromised and rotate it as soon as this commit has been released.